@@ -74,7 +74,6 @@
     "fs-extra": "7.0.1",
     "generate-changelog": "1.7.1",
     "glob": "7.1.3",
-    "highlight.js": "9.14.2",
     "html-webpack-plugin": "3.2.0",
     "inquirer": "6.3.1",
     "jest": "24.7.1",
@@ -90,7 +89,6 @@
     "start-server-webpack-plugin": "2.2.5",
     "storybook-addon-vue-info": "^0.7.1",
     "storybook-vue-router": "^1.0.3",
-    "terser": "3.14.0",
     "ts-jest": "24.0.2",
     "ts-loader": "5.4.5",
     "tslint": "5.16.0",
@@ -102,7 +100,7 @@
     "vue-server-renderer": "^2.6.10",
     "vue-ssr-webpack-plugin": "3.0.0",
     "vue-template-compiler": "^2.6.10",
-    "webpack": "4.28.4",
+    "webpack": "4.30.0",
     "webpack-bundle-analyzer": "3.3.2",
     "webpack-cli": "3.3.1",
     "webpack-dev-middleware": "3.6.2",
