@@ -1,7 +1,7 @@
 import { Command, ICommandHandler, IRunOptions } from '../lib/command';
-import { handleProcessError, runProcess } from '../utils/process';
-import { Spinner } from '../utils/ui';
+import { logError, Spinner } from '../utils/ui';
 import { Config } from '../models/Config';
+import { sync } from 'rimraf';
 
 @Command({
   name: 'clean',
@@ -9,19 +9,18 @@ import { Config } from '../models/Config';
 })
 export class Clean implements ICommandHandler {
   public async run(args: string[], options: IRunOptions) {
-    args = Config.clean.concat(args);
     const spinner = new Spinner();
 
     spinner.message = 'Cleaning directories...';
     spinner.start(options.debug);
 
     try {
-      await runProcess('rimraf', args, { silent: true, ...options });
-
+      Config.clean.forEach((glob) => sync(glob));
       spinner.message = 'Directories cleaned';
       spinner.stop();
     } catch (e) {
-      handleProcessError(e, spinner);
+      spinner.stop();
+      logError(e);
     }
   }
 }
