@@ -1,7 +1,8 @@
 import { Command, ICommandHandler, IRunOptions } from '../lib/command';
-import { handleProcessError, runProcess } from '../utils/process';
 import { packageRoot } from '../utils/path';
 import { Config } from '../models/Config';
+import { logError } from '../utils/ui';
+import { runProcess } from '../utils/process';
 
 @Command({
   name: 'generate',
@@ -11,10 +12,12 @@ import { Config } from '../models/Config';
 export class Generate implements ICommandHandler {
   public async run(args: string[], options: IRunOptions) {
     try {
-      await runProcess('plop', ['--plopfile', packageRoot('dist/generators/index.js')], options);
+      process.argv = [process.argv[0], null, '--plopfile', packageRoot('dist/generators/index.js'), ...args];
+      require('plop');
+
       await runProcess('vuesion', ['prettier', '--pattern', `${Config.generators.outputDirectory}/**/*`], options);
     } catch (e) {
-      handleProcessError(e);
+      logError(e);
     }
   }
 }
